requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
mysql-connector-python
python-dotenv